requirements.txt
Flask
flask-cors
tensorflow
Pillow
psycopg2-binary
numpy
pandas
matplotlib
scikit-image
lime
gunicorn
